Refactor SBOM management components
