+++ web/src/utils/SBOMManagement/sbomManagementUtils.js
